ivshmem: Fix compilation (wrong format specifier)
st_size is an off32_t or off64_t, so %ld does not always work.

Cc: Anthony Liguori <aliguori@us.ibm.com>
Cc: Cam Macdonell <cam@cs.ualberta.ca>

+++ hw/ivshmem.c
@@ -352,8 +352,8 @@ static int check_shm_size(IVShmemState *s, int fd) {
 
     if (s->ivshmem_size > buf.st_size) {
         fprintf(stderr, "IVSHMEM ERROR: Requested memory size greater");
-        fprintf(stderr, " than shared object size (%" PRIu64 " > %ld)\n",
-                                          s->ivshmem_size, buf.st_size);
+        fprintf(stderr, " than shared object size (%" PRIu64 " > %" PRIu64 ")\n",
+                                          s->ivshmem_size, (uint64_t)buf.st_size);
         return -1;
     } else {
         return 0;
